+++ packages/vm/src/testing/calls.rs
@@ -5,9 +5,9 @@ use schemars::JsonSchema;
 use serde::{de::DeserializeOwned, Serialize};
 use std::fmt;
 
-use cosmwasm_std::{ContractResult, Env, MessageInfo, QueryResponse, Response};
+use cosmwasm_std::{ContractResult, Env, MessageInfo, QueryResponse, Reply, Response};
 
-use crate::calls::{call_handle, call_init, call_migrate, call_query, call_system};
+use crate::calls::{call_handle, call_init, call_migrate, call_query, call_reply, call_system};
 use crate::instance::Instance;
 use crate::serde::to_vec;
 use crate::{BackendApi, Querier, Storage};
@@ -88,6 +88,23 @@ where
 {
     let serialized_msg = to_vec(&msg).expect("Testing error: Could not seralize request message");
     call_system(instance, &env, &serialized_msg).expect("VM error")
+}
+
+// reply mimicks the call signature of the smart contracts.
+// thus it moves env and msg rather than take them as reference.
+// this is inefficient here, but only used in test code
+pub fn reply<A, S, Q, U>(
+    instance: &mut Instance<A, S, Q>,
+    env: Env,
+    msg: Reply,
+) -> ContractResult<Response<U>>
+where
+    A: BackendApi + 'static,
+    S: Storage + 'static,
+    Q: Querier + 'static,
+    U: DeserializeOwned + Clone + PartialEq + JsonSchema + fmt::Debug,
+{
+    call_reply(instance, &env, &msg).expect("VM error")
 }
 
 // query mimicks the call signature of the smart contracts.